+++ open_manipulator_bringup/launch/gazebo.launch.py
@@ -17,76 +17,81 @@
 # Author: Wonho Yoon, Sungho Woo
 
 import os
-import xacro
 from pathlib import Path
-from ament_index_python.packages import get_package_share_directory
 
+from ament_index_python.packages import get_package_share_directory
 from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument, ExecuteProcess, IncludeLaunchDescription
-from launch.actions import RegisterEventHandler, SetEnvironmentVariable
+from launch.actions import DeclareLaunchArgument
+from launch.actions import ExecuteProcess
+from launch.actions import IncludeLaunchDescription
+from launch.actions import RegisterEventHandler
+from launch.actions import SetEnvironmentVariable
 from launch.event_handlers import OnProcessExit
 from launch.launch_description_sources import PythonLaunchDescriptionSource
 from launch.substitutions import LaunchConfiguration
 from launch_ros.actions import Node
-from launch_ros.substitutions import FindPackageShare
+import xacro
+
 
 def get_robot_model():
-    robot_model = os.getenv("ROBOT_MODEL", "om_y")
-    if robot_model not in ["om_x", "om_y", "om_y_follower"]:
-        raise ValueError(f"Invalid ROBOT_MODEL: {robot_model}")
+    robot_model = os.getenv('ROBOT_MODEL', 'om_y')
+    if robot_model not in ['om_x', 'om_y', 'om_y_follower']:
+        raise ValueError(f'Invalid ROBOT_MODEL: {robot_model}')
     return robot_model
+
 
 def generate_launch_description():
     # Launch Arguments
     robot_model = get_robot_model()
 
-    if robot_model == "om_x":
-        urdf_file = "open_manipulator_x"
-        urdf_folder = "om_x"
+    if robot_model == 'om_x':
+        urdf_file = 'open_manipulator_x'
+        urdf_folder = 'om_x'
     else:
-        urdf_file = "open_manipulator_y"
-        urdf_folder = "om_y"
+        urdf_file = 'open_manipulator_y'
+        urdf_folder = 'om_y'
 
     open_manipulator_description_path = os.path.join(
-        get_package_share_directory('open_manipulator_description'))
+        get_package_share_directory('open_manipulator_description')
+    )
 
     open_manipulator_bringup_path = os.path.join(
-        get_package_share_directory('open_manipulator_bringup'))
+        get_package_share_directory('open_manipulator_bringup')
+    )
 
     # Set gazebo sim resource path
     gazebo_resource_path = SetEnvironmentVariable(
         name='GZ_SIM_RESOURCE_PATH',
         value=[
-            os.path.join(open_manipulator_bringup_path, 'worlds'), ':' +
-            str(Path(open_manipulator_description_path).parent.resolve())
-            ]
-        )
-
-    arguments = LaunchDescription([
-                DeclareLaunchArgument('world', default_value='empty_world',
-                          description='Gz sim World'),
-           ]
+            os.path.join(open_manipulator_bringup_path, 'worlds'),
+            ':' + str(Path(open_manipulator_description_path).parent.resolve()),
+        ],
     )
 
+    arguments = LaunchDescription([
+        DeclareLaunchArgument(
+            'world', default_value='empty_world', description='Gz sim World'
+        ),
+    ])
+
     gazebo = IncludeLaunchDescription(
-                PythonLaunchDescriptionSource([os.path.join(
-                    get_package_share_directory('ros_gz_sim'), 'launch'), '/gz_sim.launch.py']),
-                launch_arguments=[
-                    ('gz_args', [LaunchConfiguration('world'),
-                                 '.sdf',
-                                 ' -v 1',
-                                 ' -r']
-                    )
-                ]
-             )
+        PythonLaunchDescriptionSource([
+            os.path.join(get_package_share_directory('ros_gz_sim'), 'launch'),
+            '/gz_sim.launch.py',
+        ]),
+        launch_arguments=[
+            ('gz_args', [LaunchConfiguration('world'), '.sdf', ' -v 1', ' -r'])
+        ],
+    )
 
-    xacro_file = os.path.join(open_manipulator_description_path,
-                              "urdf",
-                              urdf_folder,
-                              f"{urdf_file}.urdf.xacro")
+    xacro_file = os.path.join(
+        open_manipulator_description_path,
+        'urdf',
+        urdf_folder,
+        f'{urdf_file}.urdf.xacro',
+    )
 
-
-    doc = xacro.process_file(xacro_file, mappings={'use_sim' : 'true'})
+    doc = xacro.process_file(xacro_file, mappings={'use_sim': 'true'})
 
     robot_desc = doc.toprettyxml(indent='  ')
 
@@ -96,59 +101,90 @@ def generate_launch_description():
         package='robot_state_publisher',
         executable='robot_state_publisher',
         output='screen',
-        parameters=[params]
+        parameters=[params],
     )
 
     gz_spawn_entity = Node(
         package='ros_gz_sim',
         executable='create',
         output='screen',
-        arguments=['-string', robot_desc,
-                   '-x', '0.0',
-                   '-y', '0.0',
-                   '-z', '0.0',
-                   '-R', '0.0',
-                   '-P', '0.0',
-                   '-Y', '0.0',
-                   '-name', 'om',
-                   '-allow_renaming', 'true'
-                   '-use_sim','true'],
+        arguments=[
+            '-string',
+            robot_desc,
+            '-x',
+            '0.0',
+            '-y',
+            '0.0',
+            '-z',
+            '0.0',
+            '-R',
+            '0.0',
+            '-P',
+            '0.0',
+            '-Y',
+            '0.0',
+            '-name',
+            'om',
+            '-allow_renaming',
+            'true-use_sim',
+            'true',
+        ],
     )
 
     load_joint_state_controller = ExecuteProcess(
-        cmd=['ros2', 'control', 'load_controller', '--set-state', 'active',
-             'joint_state_broadcaster'],
-        output='screen'
+        cmd=[
+            'ros2',
+            'control',
+            'load_controller',
+            '--set-state',
+            'active',
+            'joint_state_broadcaster',
+        ],
+        output='screen',
     )
 
     load_arm_controller = ExecuteProcess(
-        cmd=['ros2', 'control', 'load_controller', '--set-state', 'active',
-             'arm_controller'],
-        output='screen'
+        cmd=[
+            'ros2',
+            'control',
+            'load_controller',
+            '--set-state',
+            'active',
+            'arm_controller',
+        ],
+        output='screen',
     )
 
     load_gripper_controller = ExecuteProcess(
-        cmd=['ros2', 'control', 'load_controller', '--set-state', 'active',
-             'gripper_controller'],
-        output='screen'
+        cmd=[
+            'ros2',
+            'control',
+            'load_controller',
+            '--set-state',
+            'active',
+            'gripper_controller',
+        ],
+        output='screen',
     )
 
     bridge = Node(
         package='ros_gz_bridge',
         executable='parameter_bridge',
         arguments=['/clock@rosgraph_msgs/msg/Clock[gz.msgs.Clock'],
-        output='screen'
+        output='screen',
     )
 
-    rviz_config_file = os.path.join(open_manipulator_description_path, 'rviz', 'open_manipulator.rviz')
+    # rviz_config_file = os.path.join(
+    #     open_manipulator_description_path, 'rviz', 'open_manipulator.rviz'
+    # )
 
-    rviz = Node(
-        package="rviz2",
-        executable="rviz2",
-        name="rviz2",
-        output="log",
-        arguments=["-d", rviz_config_file],
-    )
+    # rviz = Node(
+    #     package='rviz2',
+    #     executable='rviz2',
+    #     name='rviz2',
+    #     output='log',
+    #     arguments=['-d', rviz_config_file],
+    # )
 
     return LaunchDescription([
         RegisterEventHandler(
@@ -159,9 +195,8 @@ def generate_launch_description():
         ),
         RegisterEventHandler(
             event_handler=OnProcessExit(
-               target_action=load_joint_state_controller,
-               on_exit=[load_arm_controller,
-                        load_gripper_controller],
+                target_action=load_joint_state_controller,
+                on_exit=[load_arm_controller, load_gripper_controller],
             )
         ),
         bridge,
@@ -170,7 +205,5 @@ def generate_launch_description():
         gazebo,
         node_robot_state_publisher,
         gz_spawn_entity,
-        #rviz,
+        # rviz,
     ])
-
-