+++ open_manipulator_bringup/launch/hardware_x.launch.py
@@ -17,10 +17,15 @@
 # Author: Wonho Yoon, Sungho Woo
 
 from launch import LaunchDescription
-from launch.actions import DeclareLaunchArgument, RegisterEventHandler, TimerAction, ExecuteProcess
-from launch.conditions import IfCondition, UnlessCondition
+from launch.actions import DeclareLaunchArgument
+from launch.actions import RegisterEventHandler
+from launch.conditions import IfCondition
+from launch.conditions import UnlessCondition
 from launch.event_handlers import OnProcessExit
-from launch.substitutions import Command, FindExecutable, LaunchConfiguration, PathJoinSubstitution
+from launch.substitutions import Command
+from launch.substitutions import FindExecutable
+from launch.substitutions import LaunchConfiguration
+from launch.substitutions import PathJoinSubstitution
 from launch_ros.actions import Node
 from launch_ros.substitutions import FindPackageShare
 
@@ -28,13 +33,39 @@ from launch_ros.substitutions import FindPackageShare
 def generate_launch_description():
     # Declare launch arguments
     declared_arguments = [
-        DeclareLaunchArgument('start_rviz', default_value='true', description='Whether to execute rviz2'),
-        DeclareLaunchArgument('prefix', default_value='""', description='Prefix of the joint and link names'),
-        DeclareLaunchArgument('use_sim', default_value='false', description='Start robot in Gazebo simulation.'),
-        DeclareLaunchArgument('use_fake_hardware', default_value='false', description='Use fake hardware mirroring command.'),
-        DeclareLaunchArgument('fake_sensor_commands', default_value='false', description='Enable fake sensor commands.'),
-        DeclareLaunchArgument('port_name', default_value='/dev/ttyUSB0', description='Port name for hardware connection.'),
-        DeclareLaunchArgument('run_init_position', default_value='true', description='Run init_position_x.py after launch')
+        DeclareLaunchArgument(
+            'start_rviz', default_value='true', description='Whether to execute rviz2'
+        ),
+        DeclareLaunchArgument(
+            'prefix',
+            default_value='""',
+            description='Prefix of the joint and link names',
+        ),
+        DeclareLaunchArgument(
+            'use_sim',
+            default_value='false',
+            description='Start robot in Gazebo simulation.',
+        ),
+        DeclareLaunchArgument(
+            'use_fake_hardware',
+            default_value='false',
+            description='Use fake hardware mirroring command.',
+        ),
+        DeclareLaunchArgument(
+            'fake_sensor_commands',
+            default_value='false',
+            description='Enable fake sensor commands.',
+        ),
+        DeclareLaunchArgument(
+            'port_name',
+            default_value='/dev/ttyUSB0',
+            description='Port name for hardware connection.',
+        ),
+        DeclareLaunchArgument(
+            'run_init_position',
+            default_value='true',
+            description='Run joint_trajectory_executor after launch',
+        ),
     ]
 
     # Launch configurations
@@ -45,46 +76,72 @@ def generate_launch_description():
     fake_sensor_commands = LaunchConfiguration('fake_sensor_commands')
     port_name = LaunchConfiguration('port_name')
     run_init_position = LaunchConfiguration('run_init_position')
+    trajectory_params_file = LaunchConfiguration('trajectory_params_file')
 
     # Generate URDF file using xacro
     urdf_file = Command([
         PathJoinSubstitution([FindExecutable(name='xacro')]),
         ' ',
-        PathJoinSubstitution([FindPackageShare('open_manipulator_description'), 'urdf', 'om_x', 'open_manipulator_x.urdf.xacro']),
+        PathJoinSubstitution([
+            FindPackageShare('open_manipulator_description'),
+            'urdf',
+            'om_x',
+            'open_manipulator_x.urdf.xacro',
+        ]),
         ' ',
-        'prefix:=', prefix,
+        'prefix:=',
+        prefix,
         ' ',
-        'use_sim:=', use_sim,
+        'use_sim:=',
+        use_sim,
         ' ',
-        'use_fake_hardware:=', use_fake_hardware,
+        'use_fake_hardware:=',
+        use_fake_hardware,
         ' ',
-        'fake_sensor_commands:=', fake_sensor_commands,
+        'fake_sensor_commands:=',
+        fake_sensor_commands,
         ' ',
-        'port_name:=', port_name,
+        'port_name:=',
+        port_name,
     ])
+
+    robot_description = {'robot_description': urdf_file}
 
     # Paths for configuration files
     controller_manager_config = PathJoinSubstitution([
-        FindPackageShare('open_manipulator_bringup'), 'config', 'om_x', 'hardware_controller_manager.yaml'
+        FindPackageShare('open_manipulator_bringup'),
+        'config',
+        'om_x',
+        'hardware_controller_manager.yaml',
     ])
+
     rviz_config_file = PathJoinSubstitution([
-        FindPackageShare('open_manipulator_description'), 'rviz', 'open_manipulator.rviz'
+        FindPackageShare('open_manipulator_description'),
+        'rviz',
+        'open_manipulator.rviz',
+    ])
+
+    trajectory_params_file = PathJoinSubstitution([
+        FindPackageShare('open_manipulator_bringup'),
+        'config',
+        'om_x',
+        'initial_positions.yaml',
     ])
 
     # Define nodes
     control_node = Node(
         package='controller_manager',
         executable='ros2_control_node',
-        parameters=[{'robot_description': urdf_file}, controller_manager_config],
+        parameters=[robot_description, controller_manager_config],
         output='both',
-        condition=UnlessCondition(use_sim)
+        condition=UnlessCondition(use_sim),
     )
 
-    robot_state_pub_node = Node(
+    robot_state_publisher_node = Node(
         package='robot_state_publisher',
         executable='robot_state_publisher',
-        parameters=[{'robot_description': urdf_file, 'use_sim_time': use_sim}],
-        output='screen'
+        parameters=[robot_description, {'use_sim_time': use_sim}],
+        output='screen',
     )
 
     rviz_node = Node(
@@ -92,72 +149,51 @@ def generate_launch_description():
         executable='rviz2',
         arguments=['-d', rviz_config_file],
         output='screen',
-        condition=IfCondition(start_rviz)
+        condition=IfCondition(start_rviz),
     )
 
-    joint_state_broadcaster_spawner = Node(
+    # Controller spawner node
+    robot_controller_spawner = Node(
         package='controller_manager',
         executable='spawner',
-        arguments=['joint_state_broadcaster', '--controller-manager', '/controller_manager'],
-        output='screen'
+        arguments=[
+            'arm_controller',
+            'gripper_controller',
+            'joint_state_broadcaster',
+        ],
+        parameters=[robot_description],
     )
 
-    arm_controller_spawner = Node(
-        package='controller_manager',
-        executable='spawner',
-        arguments=['arm_controller'],
-        output='screen'
-    )
-
-    gripper_controller_spawner = Node(
-        package='controller_manager',
-        executable='spawner',
-        arguments=['gripper_controller'],
-        output='screen'
+    # Joint trajectory executor node
+    joint_trajectory_executor = Node(
+        package='open_manipulator_bringup',
+        executable='joint_trajectory_executor',
+        parameters=[trajectory_params_file],
+        output='screen',
+        condition=IfCondition(run_init_position),
     )
 
     # Event handlers to ensure order of execution
     delay_rviz_after_joint_state_broadcaster_spawner = RegisterEventHandler(
         event_handler=OnProcessExit(
-            target_action=joint_state_broadcaster_spawner,
-            on_exit=[rviz_node]
+            target_action=robot_controller_spawner, on_exit=[rviz_node]
         )
     )
 
-    delay_arm_controller_spawner_after_joint_state_broadcaster_spawner = RegisterEventHandler(
+    delay_joint_trajectory_executor_after_controllers = RegisterEventHandler(
         event_handler=OnProcessExit(
-            target_action=joint_state_broadcaster_spawner,
-            on_exit=[arm_controller_spawner]
+            target_action=robot_controller_spawner,
+            on_exit=[joint_trajectory_executor],
         )
-    )
-
-    delay_gripper_controller_spawner_after_joint_state_broadcaster_spawner = RegisterEventHandler(
-        event_handler=OnProcessExit(
-            target_action=joint_state_broadcaster_spawner,
-            on_exit=[gripper_controller_spawner]
-        )
-    )
-
-    # Timer action to run init_position.py after launch
-    init_position_timer = TimerAction(
-        period=3.0,
-        actions=[
-            ExecuteProcess(
-                cmd=['ros2', 'run', 'open_manipulator_bringup', 'init_position_x.py'],
-                output='screen',
-                condition=IfCondition(run_init_position)
-            )
-        ]
     )
 
     return LaunchDescription(
-        declared_arguments + [
+        declared_arguments
+        + [
             control_node,
-            robot_state_pub_node,
-            joint_state_broadcaster_spawner,
+            robot_controller_spawner,
+            robot_state_publisher_node,
             delay_rviz_after_joint_state_broadcaster_spawner,
-            delay_arm_controller_spawner_after_joint_state_broadcaster_spawner,
-            delay_gripper_controller_spawner_after_joint_state_broadcaster_spawner,
-            init_position_timer
+            delay_joint_trajectory_executor_after_controllers,
         ]
     )