+++ open_manipulator_gui/src/om_y/qnode.cpp
@@ -1,4 +1,4 @@
-﻿// Copyright 2024 ROBOTIS CO., LTD.
+// Copyright 2024 ROBOTIS CO., LTD.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //
-// Author: Ryan Shim, Sungho Woo
+// Author: Ryan Shim, Sungho Woo, Wonho Yoon, Woojin Wie
 
 #include "open_manipulator_y_gui/qnode.hpp"