+++ open_manipulator_moveit_config/launch/moveit_core.launch.py
@@ -18,18 +18,17 @@
 
 import os
 
+from ament_index_python.packages import get_package_share_directory
 from launch import LaunchDescription
 from launch.actions import IncludeLaunchDescription
 from launch.launch_description_sources import PythonLaunchDescriptionSource
-from ament_index_python.packages import get_package_share_directory
 
 
 def generate_launch_description():
-
     ld = LaunchDescription()
     launch_dir = os.path.join(
-        get_package_share_directory(
-            'open_manipulator_moveit_config'), 'launch')
+        get_package_share_directory('open_manipulator_moveit_config'), 'launch'
+    )
 
     # RViz
     rviz_launch = IncludeLaunchDescription(