+++ open_manipulator_description/CHANGELOG.rst
@@ -4,7 +4,8 @@ Changelog for package open_manipulator_description
 
 3.2.2 (2025-04-17)
 ------------------
-* None
+* Streamlined Ros2 control config files and Handle lint errors
+* Contributors: Woojin Wie
 
 3.2.1 (2025-04-11)
 ------------------